Start creating browser layer
